requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
pandas==2.2.1
python-dotenv==1.0.1
langchain-anthropic==0.0.2
langchain==0.1.9
pydantic==2.6.3
aiohttp==3.9.3